+++ src/main/kotlin/cloud/glitchdev/rfu/gui/hud/AbstractTextHudElement.kt
@@ -1,9 +1,7 @@
 package cloud.glitchdev.rfu.gui.hud
 
-import gg.essential.elementa.components.UIText
+import cloud.glitchdev.rfu.gui.components.elementa.MultilineHudText
 import gg.essential.elementa.constraints.CenterConstraint
-import gg.essential.elementa.constraints.ScaledTextConstraint
-import gg.essential.elementa.constraints.TextAspectConstraint
 import gg.essential.elementa.dsl.childOf
 import gg.essential.elementa.dsl.constrain
 
@@ -11,11 +9,9 @@ abstract class AbstractTextHudElement(id : String) : AbstractHudElement(id) {
     override val enabled: Boolean
         get() = super.enabled
 
-    var text : UIText = UIText().constrain {
+    var text : MultilineHudText = MultilineHudText(scale).constrain {
         x = CenterConstraint()
         y = CenterConstraint()
-        width = ScaledTextConstraint(scale)
-        height = TextAspectConstraint()
     } childOf this
 
     override fun onInitialize() {
@@ -23,8 +19,6 @@ abstract class AbstractTextHudElement(id : String) : AbstractHudElement(id) {
     }
 
     override fun onUpdateState() {
-        text.constrain {
-            width = ScaledTextConstraint(scale)
-        }
+        text.updateScale(scale)
     }
 }